change name to ExampleBackgroundCard, create basic layout for card
