+++ src/pages/SearchPage.js
@@ -8,6 +8,7 @@ import { SearchContext } from "../context/SearchContext";
 import axios from "axios";
 import Hero from "../components/Hero/Hero";
 import RelatedWordsCard from '../components/RelatedWordsCard/RelatedWordsCard';
+import ExamplesBackgroundCard from '../components/ExamplesBackgroundCard/ExamplesBackgroundCard';
 
 function SearchPage() {
   const [searchError, setSearchError] = useState("");
@@ -150,6 +151,7 @@ function SearchPage() {
       <WordCard searchError={searchError} scrabbleScore={scrabbleScore} />
       {searchedWord && <Scrabble word={searchedWord} />}
       <RelatedWordsCard />
+      <ExamplesBackgroundCard />
     </div>
   );
 }
